Added stripe to it's own component
